% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adat-helpers.R
\name{adat-helpers}
\alias{adat-helpers}
\alias{getAdatVersion}
\alias{getSomaScanVersion}
\alias{getSignalSpace}
\alias{checkSomaScanVersion}
\title{Helpers to Extract Information from an ADAT}
\usage{
getAdatVersion(atts)

getSomaScanVersion(adat)

getSignalSpace(adat)

checkSomaScanVersion(ver)
}
\arguments{
\item{atts}{The \emph{attributes} of a \code{soma_adat} object.}

\item{adat}{A \code{soma_adat} object (with intact attributes),
typically created using \code{\link[=read_adat]{read_adat()}}.}

\item{ver}{\code{character(1)}. The SomaScan version as a string.}
}
\value{
Either a character string of length 1, or \code{NULL}.
\item{\code{\link[=getAdatVersion]{getAdatVersion()}}}{The key-value of the \code{Version} as a string.}
\item{\code{\link[=getSomaScanVersion]{getSomaScanVersion()}}}{The key-value of the \code{AssayVersion} as a string.}
\item{\code{\link[=checkSomaScanVersion]{checkSomaScanVersion()}}}{Returns \code{NULL} (invisibly) if checks pass.}
}
\description{
\code{\link[=getAdatVersion]{getAdatVersion()}} determines the the ADAT version
number from a parsed ADAT header.\cr\cr
\code{\link[=getSomaScanVersion]{getSomaScanVersion()}} determines the version of
SomaScan assay that a \code{soma_adat} object was run on.
These are elements of the \code{HEADER} attributes of the object.\cr\cr
\code{\link[=checkSomaScanVersion]{checkSomaScanVersion()}} determines if the version of
is a recognized version of SomaScan.\cr
\cr
Table of current SomaScan Assay versions:
\tabular{lll}{
\strong{Version}  \tab \strong{Name} \tab \strong{Size} \cr
\code{V4}         \tab 5k       \tab 5284     \cr
\code{v4.1}       \tab 7k       \tab 7596     \cr
\code{v5.0}       \tab 11k      \tab 11083    \cr
}
}
\note{
The \code{"v"} is case insensitive.
}
\examples{
atts <- attributes(example_data)
getAdatVersion(atts)

atts$Header.Meta$HEADER$Version <- "1.0"
getAdatVersion(atts)

ver <- getSomaScanVersion(example_data)
ver

rfu_space <- getSignalSpace(example_data)
rfu_space

is.null(checkSomaScanVersion(ver))
}
\author{
Stu Field
}
