% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adat2eSet.R
\name{adat2eSet}
\alias{adat2eSet}
\title{Convert ADAT to ExpressionSet Object}
\usage{
adat2eSet(adat)
}
\arguments{
\item{adat}{A \code{soma_adat} class object as read into the R environment using
\code{\link[=read_adat]{read_adat()}} to convert to Bioconductor ExpressionSet object.}
}
\value{
An object of class \code{ExpressionSet}.
}
\description{
Utility function that converts a SomaLogic "soma_adat" object to an
"ExpressionSet" object via the \pkg{Biobase} package (download from
\code{"Bioconductor"}).
}
\details{
This function requires the \pkg{Biobase} package and must be installed from
\code{Bioconductor}. To obtain \pkg{Biobase}, type the following at the R
console:\preformatted{if (!requireNamespace("BiocManager", quietly = TRUE)) \{
  install.packages("BiocManager")
\}
BiocManager::install("Biobase")
}
}
\examples{
eSet <- adat2eSet(sample.adat)
class(eSet)
ft <- Biobase::exprs(eSet)
head(ft[, 1:10], 10)
eSet
}
\references{
https://bioconductor.org/install/
}
\author{
Stu Field
}
