% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-example-data.R
\name{create_ex_anno_tbl}
\alias{create_ex_anno_tbl}
\title{Create Example Annotation Table}
\usage{
create_ex_anno_tbl(example_data = NULL)
}
\arguments{
\item{example_data}{Optional. A \code{soma_adat} object. If \code{NULL}, will create
a full example dataset using \code{create_example_data("full")}.}
}
\value{
A tibble with analyte annotation information.
}
\description{
Returns the annotations table (column metadata) from the example dataset.
}
\examples{
anno <- create_ex_anno_tbl()
dim(anno)
names(anno)
}
\seealso{
\code{\link[=create_example_data]{create_example_data()}}, \code{\link[=getAnalyteInfo]{getAnalyteInfo()}}
}
