% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-example-data.R
\name{create_ex_clin_data}
\alias{create_ex_clin_data}
\title{Create Example Clinical Data}
\usage{
create_ex_clin_data(example_data = NULL)
}
\arguments{
\item{example_data}{Optional. A \code{soma_adat} object. If \code{NULL}, will create
a full example dataset using \code{create_example_data("full")}.}
}
\value{
A tibble with \code{SampleId}, \code{smoking_status}, and \code{alcohol_use} columns
for clinical samples.
}
\description{
Returns a clinical data table with additional sample annotations for merging
to example data.
}
\examples{
clin <- create_ex_clin_data()
dim(clin)
head(clin)
}
\seealso{
\code{\link[=create_example_data]{create_example_data()}}
}
