% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-example-data.R
\name{create_ex_target_names}
\alias{create_ex_target_names}
\title{Create Example Target Names Map}
\usage{
create_ex_target_names(example_data = NULL)
}
\arguments{
\item{example_data}{Optional. A \code{soma_adat} object. If \code{NULL}, will create
a full example dataset using \code{create_example_data("full")}.}
}
\value{
A named list of class \code{target_map} mapping SeqId feature names to
target protein names.
}
\description{
Returns a mapping of analyte names to target protein names from the
example dataset.
}
\examples{
targets <- create_ex_target_names()
length(targets)
head(targets, 3)
}
\seealso{
\code{\link[=create_example_data]{create_example_data()}}, \code{\link[=getTargetNames]{getTargetNames()}}
}
