% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffAdats.R
\name{diffAdats}
\alias{diffAdats}
\title{Diff Two ADAT Objects}
\usage{
diffAdats(adat1, adat2, tolerance = 1e-06)
}
\arguments{
\item{adat1}{First \code{soma_adat} object.}

\item{adat2}{Second \code{soma_adat} object.}

\item{tolerance}{Tolerance for the difference between
numeric vectors (i.e. SOMAmer/feature data). Passed
eventually to \code{\link[base]{all.equal}}.}
}
\value{
\code{NULL} invisibly.
}
\description{
Diff tool for the differences between two \code{soma_adat} objects.
When diffs of the table \emph{values} are interrogated, \bold{only}
the intersect of the column meta data or feature data is considered
}
\note{
Diffs of the \emph{intersect} of column names are considered.
}
\examples{
diffAdats(sample.adat, sample.adat)

# remove random column
set.seed(1)
diffAdats(sample.adat, sample.adat[, -sample(1:ncol(sample.adat), 1)])

# change TimePoint randomely
diffAdats(sample.adat, dplyr::mutate(sample.adat, TimePoint = sample(TimePoint)))

# remove Females
males  <- dplyr::filter(sample.adat, SampleGroup == "M")
mydiff <- diffAdats(sample.adat, males)      # ALL columns will now differ
}
\seealso{
\code{\link[assertthat]{are_equal}}
}
\author{
Stu Field
}
