% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{sample_frac}
\alias{sample_n}
\alias{filter}
\alias{mutate}
\alias{arrange}
\alias{ungroup}
\alias{anti_join}
\alias{full_join}
\alias{inner_join}
\alias{semi_join}
\alias{right_join}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{anti_join}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{full_join}}, \code{\link[dplyr]{inner_join}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{right_join}}, \code{\link[dplyr]{sample_frac}}, \code{\link[dplyr]{sample_n}}, \code{\link[dplyr]{semi_join}}, \code{\link[dplyr]{ungroup}}}
}}

