% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAdatVersion.R
\name{getAdatVersion}
\alias{getAdatVersion}
\title{Get ADAT Version}
\usage{
getAdatVersion(x)
}
\arguments{
\item{x}{The "Header.Meta" list of an ADAT object attributes.}
}
\value{
Returns the key-value of the ADAT version as a string.
}
\description{
Get the ADAT version number from a parsed ADAT header.
}
\examples{
header <- attributes(example_data)$Header.Meta
getAdatVersion(header)

header$HEADER$Version <- "1.0"
getAdatVersion(header)
}
\author{
Stu Field
}
