% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFeatureData.R
\name{getFeatureData}
\alias{getFeatureData}
\title{Get Feature Data}
\usage{
getFeatureData(adat)
}
\arguments{
\item{adat}{An ADAT with intact attributes (i.e. has not been modified thus
stripping original attributes), typically created using \code{\link[=read_adat]{read_adat()}}.}
}
\value{
A \code{\link[=tibble]{tibble()}} with columns corresponding to the column meta
data entries in the ADAT. The tibble has a designated column "AptName"
corresponding to the features in the ADAT, which can be retrieved using
\code{\link[=getFeatures]{getFeatures()}}. This column can be use for indexing desired analytes.
}
\description{
Uses the Column Meta data (meta data that appears above the protein
measurements in the ADAT file) from the intact attributes of a \code{soma_adat}
object and compiles them into a \code{tibble} (\code{\link[=tibble]{tibble()}}) object for simple
manipulation and indexing. The feature names of the ADAT become the
first column named \code{AptName}, which represents the key index between the
annotations table and \code{soma_adat} from which it comes.
This generates a "lookup table" that can be used for simple manipulation
and indexing of analyte annotation information.
}
\examples{
# Attribute check
is.intact.attributes(example_data)   # must be TRUE

tbl <- getFeatureData(example_data)
tbl

# Use `dplyr::group_by()`
dplyr::tally(dplyr::group_by(tbl, Dilution))     # Print summary by dilution

# Columns containing "Target"
tbl \%>\% dplyr::select(dplyr::contains("Target"))

# Rows of "Target" starting with MMP
tbl \%>\% dplyr::filter(stringr::str_detect(Target, "^MMP"))
}
\seealso{
\code{\link[=getFeatures]{getFeatures()}}
}
\author{
Stu Field
}
