% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFeatures.R
\name{getFeatures}
\alias{getFeatures}
\title{Get Features}
\usage{
getFeatures(adat, rm.controls = FALSE, n = FALSE)
}
\arguments{
\item{adat}{An ADAT object created using \code{\link[=read_adat]{read_adat()}}.
Alternatively, can be a "matrix" object with column names
corresponding to feature names, or a character vector of features}

\item{rm.controls}{Logical. Should all control and
non-human features (e.g. "NonHuman", "NonBiotins",
"Spuriomers") be removed from the returned value?}

\item{n}{Logical. Return the number of features rather
than a vector string of features.}
}
\value{
A character vector of ADAT feature ("analyte") names
or an integer number corresponding to the length of the
feature names (if \code{n = TRUE}).
}
\description{
Return the feature names (i.e. the column names for the
feature data) from a ADAT data frame, matrix, or character vector.
}
\examples{
apts <- getFeatures(sample.adat)
head(apts)
getFeatures(sample.adat, TRUE)
getFeatures(sample.adat, n = TRUE)
bb <- getFeatures(names(sample.adat))
identical(apts, bb)
}
\seealso{
\code{\link[=regex]{regex()}}, \code{\link[=grep]{grep()}}
}
\author{
Stu Field
}
