% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMeta.R
\name{getMeta}
\alias{getMeta}
\title{Get Meta Data Field Names}
\usage{
getMeta(x, n = FALSE)
}
\arguments{
\item{x}{A \code{soma_adat} object created using \code{\link[=read_adat]{read_adat()}}.
Alternatively, can be a \code{matrix} class object with column names
corresponding to feature names, or a character vector of features.}

\item{n}{Logical. Return the number of meta data field names rather
than a vector string of meta data fields?}
}
\value{
A character vector of adat meta data names or,
an integer number, the length of the meta data names.
}
\description{
Return a character vector of field names of the meta data
for an ADAT, or \code{soma.data} object.
}
\examples{
meta.vec <- getMeta(sample.adat)
head(meta.vec, 20)
getMeta(sample.adat, n = TRUE)

# test data.frame and character S3 methods
identical(getMeta(sample.adat), getMeta(names(sample.adat))) # TRUE
}
\author{
Stu Field
}
