% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{go2seqid}
\alias{go2seqid}
\title{GO Term to SeqId Mapping Object}
\source{
\url{https://geneontology.org/}
}
\description{
List containing SeqIds associated with terms from the Gene Ontology (GO)
database. Currently limited to the Biological Process and Molecular function
aspects.
}
\examples{
names(go2seqid)

# GO biological process (BP)
head(go2seqid$biological_process, n = 3)

head(go2seqid$molecular_function, n = 3)
}
\keyword{datasets}
