% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-intact-attributes.R
\name{is.intact.attributes}
\alias{is.intact.attributes}
\title{Are Attributes Intact?}
\usage{
is.intact.attributes(adat, verbose = getOption("verbose"))
}
\arguments{
\item{adat}{A \code{soma_adat} object to query.}

\item{verbose}{Logical. Should the function call be run in \emph{verbose}
mode, printing relevant diagnostic call information to the console. Defaults
to the global variable defined in \code{\link[=getOption]{getOption()}}.}
}
\value{
For \code{is.intact.attributes}: \code{TRUE} if attributes are intact,
otherwise \code{FALSE}.
}
\description{
\code{is.intact.attributes} uses a series of checks to determine
if a "soma_adat" object has a complete
set of attributes. If not, this indicates that the object has
been modified since the initial \code{\link[=read_adat]{read_adat()}} call.
Checks for the presence of both "Header.Meta" and "Col.Meta" in the
attribute names. These entries are added during the
\code{\link[=read_adat]{read_adat()}} call. Specifically, within these sections
it also checks for the presence of the following entries:
\describe{
\item{"Header.Meta" section:}{"HEADER", "COL_DATA", and "ROW_DATA"}
\item{"Col.Meta" section:}{"SeqId", "Target", "Units", and "Dilution"}
}
If any of the above they are altered or missing, \code{FALSE} is returned.
}
\examples{
# checking attributes
my_adat <- sample.adat
is.intact.attributes(my_adat)           # TRUE
is.intact.attributes(my_adat[, -303])   # doesn't break atts; TRUE
attributes(my_adat)$Col.Meta <- NULL    # break attributes
is.intact.attributes(my_adat, verbose = TRUE) # FALSE
}
\seealso{
\code{\link[=attributes]{attributes()}}
}
