% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lift-adat.R
\name{lift_adat}
\alias{lift_adat}
\alias{is_lifted}
\title{Lift an ADAT Between Assay Versions}
\usage{
lift_adat(
  adat,
  bridge = c("11k_to_7k", "11k_to_5k", "7k_to_11k", "7k_to_5k", "5k_to_11k", "5k_to_7k"),
  anno.tbl = deprecated()
)

is_lifted(adat)
}
\arguments{
\item{adat}{A \code{soma_adat} object (with intact attributes),
typically created using \code{\link[=read_adat]{read_adat()}}.}

\item{bridge}{The direction of the lift (i.e. bridge).}

\item{anno.tbl}{Deprecated.}
}
\value{
A "lifted" \code{soma_adat} object corresponding to the scaling
reference in the \code{anno.tbl}. RFU values are rounded to 1 decimal to
match standard SomaScan delivery format.

Logical. Whether \code{adat} has been lifted.
}
\description{
The SomaScan platform continually improves its technical processes
between assay versions; from changing reagents, liquid handling equipment,
and increased analyte content. However, these upgrades can result in
minute differences in RFU space for a given analyte, requiring a calibration
(aka "lifting" or "bridging") to bring RFUs into a comparable space.
This is accomplished by applying an analyte-specific scalar to each analyte
RFU (ADAT column). The scalar values themselves are typically provided
via \verb{*.xlsx} file, which can be parsed via \code{\link[=read_annotations]{read_annotations()}}. See Details.

\code{\link[=is_lifted]{is_lifted()}} checks whether an object
has been lifted (bridged) by the presence
(or absence) of the \code{SignalSpace} entry
in the \code{soma_adat} attributes.
}
\details{
Lifting between various versions requires a specific
annotations file containing scalars specific to desired lifting direction.
For example, to "lift" between \code{v4.1} -> \code{v4.0}, you \emph{must} be working
with SomaScan data in \code{v4.1} space and an annotations file containing
scalars to convert \strong{to} \code{v4.0}.
Likewise, "lifting" from \code{v4.0} -> \code{v4.1} requires
a separate annotations file and a \code{soma_adat} from SomaScan \code{v4.0}.
}
\examples{
# `example_data` is SomaScan V4 (5k)
adat <- head(example_data, 3L)
getSomaScanVersion(example_data)

# perform the 'lifting'
lift_7k <- lift_adat(adat, "5k_to_7k")
is_lifted(lift_7k)

# attributes updated to reflect the 'lift'
attr(lift_7k, "Header")$HEADER$SignalSpace
attr(lift_7k, "Header")$HEADER$ProcessSteps
}
