% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MathGenerics.R
\name{MathGenerics}
\alias{MathGenerics}
\alias{Math.soma_adat}
\title{Mathematical Group Generics for ADAT Object}
\usage{
\method{Math}{soma_adat}(x, ...)
}
\arguments{
\item{x}{The \code{soma_adat} class object to perform the transformation.}

\item{...}{Additional arguments passed to the various group generics
as appropriate.}
}
\value{
A \code{soma_adat} object with the same dimensions of the input
object with the feature columns transformed by the specified generic.
}
\description{
This is the S3 group generic method to apply mathematical functions
to the RFU data of \code{soma_adat} objects.
The clinical meta data is \emph{not} transformed and remains in
the returned object. Typical generic functions include:\cr
\itemize{
\item log
\item abs
\item sign
\item floor
\item sqrt
\item exp
\item See \code{\link[=groupGeneric]{groupGeneric()}} (\emph{Math}) for full listing
}
}
\examples{
sample.adat$ACY1.3343.1.4

# log-transformation
a <- log(sample.adat)
a$ACY1.3343.1.4
b <- log10(sample.adat)
b$ACY1.3343.1.4
isTRUE(all.equal(b, log(sample.adat, base = 10)))

# floor
c <- floor(sample.adat)
c$ACY1.3343.1.4

# square-root
d <- sqrt(sample.adat)
d$ACY1.3343.1.4

# rounding
e <- round(sample.adat)
e$ACY1.3343.1.4
}
\seealso{
\code{\link[=groupGeneric]{groupGeneric()}}
}
\author{
Stu Field
}
