% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msig2seqid}
\alias{msig2seqid}
\title{MSigDB Gene Set to SeqId Mapping Object}
\source{
\url{https://www.gsea-msigdb.org/gsea/msigdb}
}
\description{
List containing SeqIds associated with gene sets from the Molecular
Signatures Database (MSigDB).
}
\examples{
names(msig2seqid)

# GO biological process (BP)
head(msig2seqid$MSigDB_oncogenic_signature, n = 3)
}
\keyword{datasets}
