% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path2seqid.R
\name{path2seqid}
\alias{path2seqid}
\title{Convert Pathway Term to SeqIds}
\usage{
path2seqid(path, format = c("go", "msigdb"))
}
\arguments{
\item{path}{Character. A pathway identifier or vector of pathway identifiers.
Pathway identifiers must be from the GO or MSigDB data resources.}

\item{format}{Character. The database to which the value in \code{path} belongs.
Accepted values are "GO" or "MSigDB" (case insensitive).}
}
\value{
A list containing a vector of SeqIds. If multiple identifiers are
provided to \code{path}, the returned list will contain a sub-list for each
identifier.
}
\description{
Utility function to convert a GO term or other biological/functional pathway
to a vector of SeqIds associated with the features (most commonly genes) in
that pathway.
}
\examples{
# Single GO term
path2seqid("GO:0007596")

# Vector of GO terms
path2seqid(c("GO:0007596", "GO:0098927", "GO:0002065"))

# MSigDB gene set
path2seqid("M2622_GLI1_UP.V1_DN", format = "msigdb")

# Pathways are case-insensitive and can be used as keywords for a search
path2seqid("myeloid", format = "msigdb") |> head(n = 2)
}
\author{
Amanda Hiser
}
