% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivotExpressionSet.R
\name{pivotExpressionSet}
\alias{pivotExpressionSet}
\title{Convert to Long Format}
\usage{
pivotExpressionSet(eSet)
}
\arguments{
\item{eSet}{An \code{ExpressionSet} class object, created using \code{\link[=adat2eSet]{adat2eSet()}}.}
}
\value{
A \code{tibble} consisting of the long format
conversion of an \code{ExpressionSet} object.
}
\description{
Utility to convert an \code{"ExpressionSet"} object
from the "wide" data format to the "long" format via \code{\link[=pivot_longer]{pivot_longer()}}.
The \pkg{Biobase} package is required for this function.
}
\examples{
# subet into a reduced mini-ADAT object
# 10 samples (rows)
# 5 clinical variables and 3 features (cols)
sub_adat <- example_data[1:10, c(1:5, 35:37)]
ex_set   <- adat2eSet(sub_adat)

# convert ExpressionSet object to long format
adat_long <- pivotExpressionSet(ex_set)
}
\seealso{
\code{\link[=adat2eSet]{adat2eSet()}}
}
\author{
Stu Field
}
