% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcessAdat.R
\name{preProcessAdat}
\alias{preProcessAdat}
\title{Pre-Process an ADAT Object for Analysis}
\usage{
preProcessAdat(
  adat,
  filter.features = TRUE,
  filter.samples = TRUE,
  data.qc = NULL,
  log.10 = TRUE,
  center.scale = TRUE
)
}
\arguments{
\item{adat}{A \code{soma_adat} object created using \code{\link[=read_adat]{read_adat()}}, including
SeqId columns (\code{seq.xxxxx.xx}) containing raw RFU values.}

\item{filter.features}{Logical. Should non-human protein features be
dropped? Default is \code{TRUE}.}

\item{filter.samples}{Logical. Should buffer, calibrator, and QC samples be
dropped, along with samples that do not pass default normalization acceptance
criteria or are identified as sample level RFU outliers?  Default is \code{TRUE}.}

\item{data.qc}{Character. Character vector of variable names for which data
QC plots are desired. Default is \code{NULL}, which does not generate any QC
plots.  Note: These plots are for visual inspection only, no samples or
features are dropped from the output \code{soma_adat} object.}

\item{log.10}{Logical. Should the RFU values be log10 transformed?
Default is \code{TRUE}.}

\item{center.scale}{Logical. Should the RFU values be Z-transformed
(centered and scaled)? Default is \code{TRUE}.}
}
\value{
A \code{soma_adat} object.
}
\description{
Pre-process an ADAT file containing raw analyte RFU values in preparation
for analysis.
}
\details{
The \code{soma_adat} object is pre-processed with the following steps:
\enumerate{
\item Filter features -> down to human protein analytes
\item Filter samples -> down to clinical samples (dropping buffer, calibrator, and
QC samples) that pass default normalization acceptance criteria, and
dropping sample-level outliers by RFU values.
\item Data QC -> plots of normalization scale factors by clinical covariates
\item Transformations -> log10, center, and scale analyte RFU values
}
}
\examples{
preProcessAdat(example_data, data.qc = c("Age", "Sex"))
}
\author{
Caleb Scheidel
}
