% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-annotations.R
\name{read_annotations}
\alias{read_annotations}
\title{Import a SomaLogic Annotations File}
\usage{
read_annotations(file)
}
\arguments{
\item{file}{A path to an annotations file location.
This should be a SomaLogic annotations file in
\verb{*.xlsx} format.}
}
\value{
A \code{tibble} containing analyte-specific annotations and
related (e.g. lift/bridging) information, keyed on SomaLogic
\link{SeqId}, the unique SomaScan analyte identifier.
}
\description{
Import a SomaLogic Annotations File
}
\examples{
\dontrun{
  # for example
  file <- "~/Downloads/SomaScan_11K_v5.0_Plasma_Serum_Annotated_Menu.xlsx"
  anno_tbl <- read_annotations(file)
}
}
