% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAnnotations.R
\name{readAnnotations}
\alias{readAnnotations}
\title{Import a SomaLogic Annotations File}
\usage{
readAnnotations(file)
}
\arguments{
\item{file}{A path to an annotations file location. This is a sanctioned,
versioned file provided by SomaLogic, Inc. and should be an unmodified
\verb{*.xlsx} file.}
}
\value{
A \code{tibble} containing analyte-specific annotations and
related information (e.g. lift/scale information), keyed on SomaLogic
\code{"SeqId"} which is a unique analyte identifier.
}
\description{
Import a SomaLogic Annotations File
}
\examples{
\dontrun{
anno <- readAnnotations("~/Desktop/SomaScan_V4.1_7K_Annotated_Content_20210616.xlsx")
}
}
