% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-reexports.R, R/tidyr-reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{count}
\alias{rename}
\alias{slice_sample}
\alias{slice}
\alias{sample_frac}
\alias{sample_n}
\alias{filter}
\alias{mutate}
\alias{arrange}
\alias{group_by}
\alias{ungroup}
\alias{left_join}
\alias{anti_join}
\alias{full_join}
\alias{inner_join}
\alias{semi_join}
\alias{right_join}
\alias{unite}
\alias{separate}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:filter-joins]{anti_join}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{count}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr:mutate-joins]{full_join}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr:mutate-joins]{inner_join}}, \code{\link[dplyr:mutate-joins]{left_join}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr:mutate-joins]{right_join}}, \code{\link[dplyr:sample_n]{sample_frac}}, \code{\link[dplyr]{sample_n}}, \code{\link[dplyr:filter-joins]{semi_join}}, \code{\link[dplyr]{slice}}, \code{\link[dplyr:slice]{slice_sample}}, \code{\link[dplyr:group_by]{ungroup}}}

  \item{tidyr}{\code{\link[tidyr]{separate}}, \code{\link[tidyr]{unite}}}
}}

