% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-gsea.R
\name{run_gsea}
\alias{run_gsea}
\title{Perform Gene Set Enrichment Analysis (GSEA) with SomaScan}
\usage{
run_gsea(ranks, pathIds, adat, min_feats = 15L, max_feats = NULL, nperm = 1000)
}
\arguments{
\item{ranks}{A two column \code{data.frame} of statistical test results to be
used as ranks. The data frame must include the following columns:
\describe{
\item{SeqId}{Character. SomaScan \code{SeqIds} or R-compatible
\code{AptNames}. For more information about these identifier formats, please
see \link{SeqId}.}
\item{Rank}{Numeric. Ranking values associated with each identifier.}
}
The suggested value for the \code{Rank} column is the absolute value of a test
statistic, typically log2 fold-change values. However, \emph{any} prior
statistical test result can be used, as long as the data frame contains
SomaScan identifiers and a numeric metric that can be used for ranking.

Alternatively, a named numeric vector can also be used. The vector must
contain ranking values associated with each vector. Names must be SomaScan
identifiers.}

\item{pathIds}{List containing character vectors of SomaScan identifiers
associated with pathways, gene sets, or functional groups of interest. The
name of each list component should be the name of the pathway/gene set. The
elements in each character vector should be SomaScan \code{SeqIds} or \code{AptNames}.
See the \link{go2seqid} object for reference.}

\item{adat}{Original \code{soma_adat} object used to obtain statistical test
results in \code{ranks}.}

\item{min_feats}{Numeric. Minimum number of features within a feature set
(i.e. vectors of \code{SeqIds} in the list provided for \code{pathIds}) for that set
to be retained. Default is 15. Sets smaller than this value will be discarded.
If \code{NULL}, no minimum is used. See \code{Details} for more information.}

\item{max_feats}{Numeric. Maximum number of features within a category
for that category to be retained. Default is \code{NULL}, meaning that no
upper bound will be used. See \code{Details} for more information.}

\item{nperm}{Numeric. Number of permutations to perform. Default is 1000.}
}
\value{
A \code{data.frame} object containing the results of gene set enrichment,
including the following columns:
\item{pathway}{Name of the pathway, as in \code{names(pathIds)}}
\item{pval}{Enrichment p-value}
\item{padj}{FDR-adjusted p-value}
\item{log2err}{The expected error for the standard deviation of the p-value logarithm}
\item{ES}{Enrichment score, calculated via the \href{https://www.gsea-msigdb.org/gsea/doc/GSEAUserGuideTEXT.htm#_Enrichment_Score_(ES)}{Broad GSEA implementation}}
\item{NES}{Enrichment score normalized to mean enrichment of random samples of the same size}
\item{Size}{Size of the pathway \emph{after} removing \code{SeqIds} not present in \code{ranks} (this step is performed before GSEA)}
\item{leadingEdge}{Vector with indexes of leading edge genes that drive the enrichment, see \href{http://software.broadinstitute.org/gsea/doc/GSEAUserGuideTEXT.htm#_Running_a_Leading}{Running a Leading Edge Analysis}}
}
\description{
Implements the algorithm from the \link{fgsea} package to perform gene set
enrichment (GSEA), tailored specifically to SomaScan data.
}
\section{Min/Max Set Sizes}{

GSEA normalization is not accurate for very small feature sets - for
example, for sets with >10 features, as few as 2-3 features can produce
significant results. To avoid this, any feature sets smaller than
\code{min_feats} will be discarded before performing GSEA. Please note that this threshold
is applied \emph{after} filtering to remove features (i.e. SomaScan identifiers)
from the feature set (gene list or other functional group) that are not
present in the \code{ranks} data set. The default value (N = 15) is based on
the minimum gene set threshold used in the Broad Institute's
\href{https://www.gsea-msigdb.org/gsea/index.jsp}{GSEA software}.
}

\examples{
ranks <- t_tests[,c("SeqId", "rank")]
results <- run_gsea(ranks = ranks, pathIds = go2seqid$biological_process,
                    example_data)
head(results)
}
\references{
A. Subramanian, P. Tamayo, V.K. Mootha, S. Mukherjee, B.L. Ebert,
M.A. Gillette, A. Paulovich, S.L. Pomeroy, T.R. Golub, E.S. Lander, &
J.P. Mesirov, Gene set enrichment analysis: A knowledge-based approach for
interpreting genome-wide expression profiles,
Proc. Natl. Acad. Sci. U.S.A. 102 (43) 15545-15550,
https://doi.org/10.1073/pnas.0506580102 (2005).
}
\author{
Amanda Hiser
}
