% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-soma-adat.R, R/s3-print-soma-adat.R,
%   R/s3-summary-soma-adat.R
\name{soma_adat}
\alias{soma_adat}
\alias{print.soma_adat}
\alias{summary.soma_adat}
\alias{[.soma_adat}
\alias{$.soma_adat}
\alias{[[.soma_adat}
\alias{[<-.soma_adat}
\alias{$<-.soma_adat}
\alias{[[<-.soma_adat}
\alias{median.soma_adat}
\title{The \code{soma_adat}Class and S3 Methods}
\usage{
\method{print}{soma_adat}(x, show_header = FALSE, ...)

\method{summary}{soma_adat}(object, tbl = NULL, digits = max(3L, getOption("digits") - 3L), ...)

\method{[}{soma_adat}(x, i, j, drop = TRUE)

\method{$}{soma_adat}(x, name)

\method{[[}{soma_adat}(x, i, j, ..., exact = TRUE)

\method{[}{soma_adat}(x, i, j, ...) <- value

\method{$}{soma_adat}(x, i, j, ...) <- value

\method{[[}{soma_adat}(x, i, j, ...) <- value

\method{median}{soma_adat}(x, na.rm = FALSE, ...)
}
\arguments{
\item{x, object}{A \code{soma_adat} class object.}

\item{show_header}{Logical. Should all the \verb{Header Data} information
be displayed instead of the data frame (\code{tibble}) object?}

\item{...}{potentially further arguments for methods; not used in
    the default method.}

\item{tbl}{An annotations table. If \code{NULL} (default),
annotation information is extracted from the object itself (if possible).
Alternatively, the result of a call to \code{\link[=getAnalyteInfo]{getAnalyteInfo()}}, from
which Target names can be extracted.}

\item{digits}{Integer. Used for number formatting with \code{\link[=signif]{signif()}}.}

\item{i}{
    indices specifying elements to extract or replace.  Indices are
    \code{numeric} or \code{character} vectors or empty (missing) or
    \code{NULL}.  Numeric values are coerced to integer as by
    \code{\link[base]{as.integer}} (and hence truncated towards zero).
    Character vectors will be matched to the \code{\link[base]{names}} of the
    object (or for matrices/arrays, the \code{\link[base]{dimnames}}):
    see \sQuote{Character indices} below for further details.

    For \code{[}-indexing only: \code{i}, \code{j}, \code{\dots} can be
    logical vectors, indicating elements/slices to select.  Such vectors
    are recycled if necessary to match the corresponding extent.
    \code{i}, \code{j}, \code{\dots} can also be negative integers,
    indicating elements/slices to leave out of the selection.

    When indexing arrays by \code{[} a single argument \code{i} can be a
    matrix with as many columns as there are dimensions of \code{x}; the
    result is then a vector with elements corresponding to the sets of
    indices in each row of \code{i}.

    An index value of \code{NULL} is treated as if it were \code{integer(0)}.
  }

\item{j}{
    indices specifying elements to extract or replace.  Indices are
    \code{numeric} or \code{character} vectors or empty (missing) or
    \code{NULL}.  Numeric values are coerced to integer as by
    \code{\link[base]{as.integer}} (and hence truncated towards zero).
    Character vectors will be matched to the \code{\link[base]{names}} of the
    object (or for matrices/arrays, the \code{\link[base]{dimnames}}):
    see \sQuote{Character indices} below for further details.

    For \code{[}-indexing only: \code{i}, \code{j}, \code{\dots} can be
    logical vectors, indicating elements/slices to select.  Such vectors
    are recycled if necessary to match the corresponding extent.
    \code{i}, \code{j}, \code{\dots} can also be negative integers,
    indicating elements/slices to leave out of the selection.

    When indexing arrays by \code{[} a single argument \code{i} can be a
    matrix with as many columns as there are dimensions of \code{x}; the
    result is then a vector with elements corresponding to the sets of
    indices in each row of \code{i}.

    An index value of \code{NULL} is treated as if it were \code{integer(0)}.
  }

\item{drop}{For matrices and arrays.  If \code{TRUE} the result is
    coerced to the lowest possible dimension (see the examples).  This
    only works for extracting elements, not for the replacement.  See
    \code{\link[base]{drop}} for further details.
  }

\item{name}{
    A literal character string or a \link[base]{name} (possibly \link[base]{backtick}
    quoted).  For extraction, this is normally (see under
    \sQuote{Environments}) partially matched to the \code{\link[base]{names}}
    of the object.
  }

\item{exact}{Ignored with a \code{\link[=warning]{warning()}}.}

\item{value}{typically an array-like \R object of a similar class as
    \code{x}.}

\item{na.rm}{a logical value indicating whether \code{NA}
    values should be stripped before the computation proceeds.}
}
\description{
The \code{soma_adat} data structure is the primary internal \code{R} representation
of SomaScan data. A \code{soma_adat} is automatically created via \code{\link[=read_adat]{read_adat()}}
when loading a \verb{*.adat} text file. It consists of a \code{data.frame}-like
object with leading columns as clinical variables and SomaScan RFU data
as the remaining variables. Two main attributes corresponding to analyte
and SomaScan run information contained in the \verb{*.adat} file are added:
\itemize{
\itemize{
\item \code{Header.Meta}: information about the SomaScan run, see \code{\link[=parseHeader]{parseHeader()}}
or \code{attr(x, "Header.Meta")}
\item \code{Col.Meta}: annotations information about the SomaScan reagents/analytes,
see \code{\link[=getAnalyteInfo]{getAnalyteInfo()}} or \code{attr(x, "Col.Meta")}
\item \code{file_specs}: parsing specifications for the ingested \verb{*.adat} file
\item \code{row_meta}: the names of the non-RFU fields. See \code{\link[=getMeta]{getMeta()}}.
}
See \code{\link[=groupGenerics]{groupGenerics()}} for a details on \code{\link[=Math]{Math()}}, \code{\link[=Ops]{Ops()}}, and \code{\link[=Summary]{Summary()}}
methods that dispatch on class \code{soma_adat}.
\cr\cr
See \code{\link[=reexports]{reexports()}} for a details on re-exported S3 generics from other
packages (mostly \code{dplyr} and \code{tidyr}) to enable S3 methods to be
dispatched on class \code{soma_adat}.
\cr\cr
Below is a list of \emph{all} currently available S3 methods that dispatch on
the \code{soma_adat} class:
}\preformatted{##  [1] [            [[           [[<-         [<-          ==          
##  [6] $            $<-          anti_join    arrange      count       
## [11] filter       full_join    getAnalytes  getMeta      group_by    
## [16] inner_join   is_seqFormat left_join    Math         median      
## [21] mutate       Ops          print        rename       right_join  
## [26] sample_frac  sample_n     select       semi_join    separate    
## [31] slice_sample slice        summary      Summary      transform   
## [36] ungroup      unite       
## see '?methods' for accessing help and source code
}

S3 \code{\link[=print]{print()}} method returns summary information parsed from the object
attributes, if present, followed by a dispatch to the \code{\link[=tibble]{tibble()}} print method.
Rownames are printed as the first column in the print method only.

S3 \code{\link[=summary]{summary()}} method returns the following for each column of the ADAT
object containing SOMAmer data (clinical meta data is \emph{excluded}):
\itemize{
\item Target (if available)
\item Minimum value
\item 1st Quantile
\item Median
\item Mean
\item 3rd Quantile
\item Maximum value
\item Standard deviation
\item Median absolute deviation (\code{\link[=mad]{mad()}})
\item Interquartile range (\code{\link[=IQR]{IQR()}})
}

S3 \code{\link[=Extract]{Extract()}} method is used for sub-setting a \code{soma_adat}
object and relies heavily on the \code{[} method that maintains the \code{soma_adat}
attributes intact \emph{and} subsets the \code{Col.Meta} so that it is consistent
with the newly created object.

S3 extraction via \code{$} is fully supported, however, as opposed to the
\code{data.frame} method, partial matching is \emph{not} allowed for class \code{soma_adat}.

S3 extraction via \code{[[} is fully supported, however, we restrict
the usage of \code{[[} for \code{soma_adat}. Use only a numeric index (e.g. \code{1L})
or a character identifying the column (e.g. \code{"SampleID"}).
Do not use \verb{[[i,j]]} syntax with \code{[[}, use \code{[} instead.
As with \code{$}, partial matching is \emph{not} allowed.

S3 assignment via \code{[} is fully supported for class \code{soma_adat}.

S3 assignment via \code{$} is fully supported for class \code{soma_adat}.

S3 assignment via \code{[[} is fully supported for class \code{soma_adat}.

S3 \code{\link[=median]{median()}} is currently unavailable for the \code{soma_adat} class.
}
\examples{
my_adat <- system.file("example", "example_data.adat",
                       package = "SomaDataIO", mustWork = TRUE) |> read_adat()
# S3 print method
my_adat

# S3 summary method
# MMP (4) analytes
mmps <- c("seq.2579.17", "seq.2788.55", "seq.2789.26", "seq.4925.54")
summary(my_adat[, mmps])

# Summarize by group
my_adat[, mmps] |>
  split(my_adat$Sex) |>
  lapply(summary)

# Alternatively pass annotations with Target info
anno <- getAnalyteInfo(my_adat)
summary(my_adat[, mmps], tbl = anno)
}
\seealso{
\code{\link[=groupGenerics]{groupGenerics()}}

Other IO: 
\code{\link{loadAdatsAsList}()},
\code{\link{parseHeader}()},
\code{\link{read_adat}()},
\code{\link{write_adat}()}
}
\concept{IO}
