% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SomaDataIO.R
\docType{package}
\name{SomaDataIO-package}
\alias{SomaDataIO}
\alias{SomaDataIO-package}
\title{SomaDataIO: SomaData Input/Output Package}
\description{
The SomaDataIO package loads the SomaLogic, Inc.
    proprietary data file, called an ADAT file ('*.adat'). It also
    provides auxiliary functions for extracting and manipulating relevant
    information from an ADAT, as well as exporting a modified ADAT to
    file.
}
\details{
To load an ADAT file into the global workspace, call
\code{read_adat("path/to/file.adat")}. This function parses the main data
table into a \code{data.frame} and assigns the remaining data from the
\code{"*.adat"} file as object \code{attributes}, call \code{attributes(adat)}.
The other functions in the package are designed to make extracting data
from and manipulating the newly created \code{soma_adat} object convenient.

Those familiar with micro-array data analysis and associated packages, e.g.
\pkg{Biobase}, will notice that the feature data (proteins) are arranged as
columns and the samples (arrays) are the rows of the data frame. This is the
inverse of typical micro-array data. This conflict can be easily solved
using the transpose function, \code{t()}, which is part of the
"package:base" in a standard R installation. In addition, those familiar
with the standard \code{ExpressionSet} object, available from
\code{Bioconductor}, might find the functions \code{\link[=adat2eSet]{adat2eSet()}} and
\code{\link[=pivotExpressionSet]{pivotExpressionSet()}} particularly useful.
}
\examples{
# For a listing of all SomaDataIO functions
library(help = SomaDataIO)

# To load an original adat call read_adat():
# There is a sample *.adat file provided in SomaDataIO
file <- system.file("sample", "sample.adat",
                    package = "SomaDataIO", mustWork = TRUE)
my_adat <- read_adat(file)

# Object class `soma_adat`
class(my_adat)
is.soma_adat(my_adat)

# Annotations Lookup Table
feature_table <- getFeatureData(my_adat)
feature_table

# Find all analytes starting with "MMP" in `feature_table`
feature_table \%>\%
  dplyr::filter(grepl("^MMP", Target))
}
\seealso{
Useful links:
\itemize{
  \item \url{www.somalogic.com}
}

}
\author{
\strong{Maintainer}: Stu Field \email{sfield@somalogic.com}

Other contributors:
\itemize{
  \item SomaLogic, Inc. [copyright holder, funder]
}

}
\keyword{internal}
\keyword{package}
