% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{t_tests}
\alias{t_tests}
\title{t-test Example Data}
\source{
Standard BioTools, Inc.
}
\description{
Results of performing a t-test with an example SomaScan dataset.
See the SomaDataIO "Two-Group Comparison" workflow for details.
}
\examples{
head(t_tests)
}
\keyword{datasets}
