% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-adat.R
\name{write_adat}
\alias{write_adat}
\title{Write an ADAT to File}
\usage{
write_adat(x, file)
}
\arguments{
\item{x}{An object of class \code{"soma_adat"}. Both \code{\link[=is.soma_adat]{is.soma_adat()}} and
\code{\link[=is.intact.attributes]{is.intact.attributes()}} must be \code{TRUE}.}

\item{file}{Character. File path where the object should be written.
For example, extensions should be \verb{*.adat}.}
}
\description{
One can write an existing modified internal ADAT
(\code{soma_adat} R object) to an external file.
However the ADAT object itself \emph{must} have intact
attributes, see \code{\link[=is.intact.attributes]{is.intact.attributes()}}.
}
\details{
The ADAT specification \emph{no longer} requires Windows
end of line (EOL) characters (\verb{"\r\n"}).
The EOL is currently \verb{"\n"} which is commonly used in POSIX systems,
like MacOS and Linux.
The EOL affects the format of the resulting file, particularly
calculating a checksum, therefore ADATs written via other systems may
result in differing EOLs. EOL encoding for operating systems is below:\cr
\tabular{llc}{
Symbol \tab Platform    \tab Character \cr
LF     \tab Linux       \tab \verb{"\n"} \cr
CR     \tab MacOS       \tab \verb{"\r"} \cr
CRLF   \tab DOS/Windows \tab \verb{"\r\n"}
}
}
\seealso{
\code{\link[=read_adat]{read_adat()}}, \code{\link[=write_tsv]{write_tsv()}}, \code{\link[=is.intact.attributes]{is.intact.attributes()}}

Other IO: 
\code{\link{parseHeader}()},
\code{\link{read_adat}()}
}
\author{
Stu Field
}
\concept{IO}
