% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adat-helpers.R, R/getSomaScanLiftCCC.R
\name{adat-helpers}
\alias{adat-helpers}
\alias{getAdatVersion}
\alias{getSomaScanVersion}
\alias{getSignalSpace}
\alias{checkSomaScanVersion}
\alias{getSomaScanLiftCCC}
\title{Helpers to Extract Information from an ADAT}
\usage{
getAdatVersion(atts)

getSomaScanVersion(adat)

getSignalSpace(adat)

checkSomaScanVersion(ver)

getSomaScanLiftCCC(matrix = c("plasma", "serum"))
}
\arguments{
\item{atts}{The \emph{attributes} of a \code{soma_adat} object.}

\item{adat}{A \code{soma_adat} object (with intact attributes),
typically created using \code{\link[=read_adat]{read_adat()}}.}

\item{ver}{\code{character(1)}. The SomaScan version as a string.
\strong{Note:} the \code{"v"}-prefix is case \emph{in}sensitive.}

\item{matrix}{Character. A string of (usually) either
\code{"serum"} or \code{"plasma"}.}
}
\value{
\item{\code{\link[=getAdatVersion]{getAdatVersion()}}}{The key-value of the \code{Version} as a string.}
\item{\code{\link[=getSomaScanVersion]{getSomaScanVersion()}}}{The key-value of the \code{AssayVersion} as a string.}
\item{\code{\link[=checkSomaScanVersion]{checkSomaScanVersion()}}}{Returns \code{NULL} (invisibly) if checks pass.}

\item{\code{\link[=getSomaScanLiftCCC]{getSomaScanLiftCCC()}}}{Returns a tibble of either the
\code{serum} or \code{plasma} CCC between various versions of the SomaScan assay.}
}
\description{
\code{\link[=getAdatVersion]{getAdatVersion()}} determines the the ADAT version
number from a parsed ADAT header.\cr\cr
\code{\link[=getSomaScanVersion]{getSomaScanVersion()}} determines the version of
SomaScan assay that a \code{soma_adat} object was run on.
These are elements of the \code{HEADER} attributes of the object.\cr\cr
\code{\link[=checkSomaScanVersion]{checkSomaScanVersion()}} determines if the version of
is a recognized version of SomaScan.\cr
\cr
Table of SomaScan Assay versions:
\tabular{lll}{
\strong{Version}  \tab \strong{Commercial Name} \tab \strong{Size} \cr
\code{V4}         \tab 5k                  \tab 5284     \cr
\code{v4.1}       \tab 7k                  \tab 7596     \cr
\code{v5.0}       \tab 11k                 \tab 11083    \cr
}

\code{\link[=getSomaScanLiftCCC]{getSomaScanLiftCCC()}} accesses the lifting Concordance Correlation
Coefficients between various SomaScan versions. For more about
CCC metrics see \code{\link[=lift_adat]{lift_adat()}}.
}
\examples{
atts <- attributes(example_data)
getAdatVersion(atts)

atts$Header.Meta$HEADER$Version <- "1.0"
getAdatVersion(atts)

ver <- getSomaScanVersion(example_data)
ver

rfu_space <- getSignalSpace(example_data)
rfu_space

is.null(checkSomaScanVersion(ver))

# plasma (default)
getSomaScanLiftCCC()

# serum
getSomaScanLiftCCC("serum")
}
\references{
Lin, Lawrence I-Kuei. 1989. A Concordance Correlation
Coefficient to Evaluate Reproducibility. \strong{Biometrics}. 45:255-268.
}
\author{
Stu Field
}
