% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adat2eSet.R
\name{adat2eSet}
\alias{adat2eSet}
\title{Convert ADAT to ExpressionSet Object}
\usage{
adat2eSet(adat)
}
\arguments{
\item{adat}{A \code{soma_adat} class object as read into the R
environment using \code{\link[=read_adat]{read_adat()}}.}
}
\value{
A Bioconductor object of class \code{ExpressionSet}.
}
\description{
Utility to convert a SomaLogic \code{soma_adat} object to an
\code{ExpressionSet} object via the \pkg{Biobase} package
(download from \code{"Bioconductor"}).
}
\details{
The \pkg{Biobase} package is required and must be installed from
\code{Bioconductor} via the following at the R console:\preformatted{if (!requireNamespace("BiocManager", quietly = TRUE)) \{
  install.packages("BiocManager")
\}
BiocManager::install("Biobase")
}
}
\examples{
eSet <- adat2eSet(example_data)
class(eSet)
eSet

ft <- Biobase::exprs(eSet)
head(ft[, 1:10], 10)
}
\references{
https://bioconductor.org/install/
}
\seealso{
\code{\link[=pivotExpressionSet]{pivotExpressionSet()}}
}
\author{
Stu Field
}
