% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-example-data.R
\name{create_ex_analytes}
\alias{create_ex_analytes}
\title{Create Example Analytes Vector}
\usage{
create_ex_analytes(example_data = NULL)
}
\arguments{
\item{example_data}{Optional. A \code{soma_adat} object. If \code{NULL}, will create
a full example dataset using \code{create_example_data("full")}.}
}
\value{
A character vector of analyte names (\code{seq.XXXX.XX} format).
}
\description{
Returns the analyte (feature) variable names from the example dataset.
}
\examples{
apts <- create_ex_analytes()
head(apts)
length(apts)
}
\seealso{
\code{\link[=create_example_data]{create_example_data()}}, \code{\link[=getAnalytes]{getAnalytes()}}
}
