% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffAdats.R
\name{diffAdats}
\alias{diffAdats}
\title{Diff Two ADAT Objects}
\usage{
diffAdats(adat1, adat2, tolerance = 1e-06)
}
\arguments{
\item{adat1}{First \code{soma_adat} object.}

\item{adat2}{Second \code{soma_adat} object.}

\item{tolerance}{Tolerance for the difference between
numeric vectors (i.e. SOMAmer/feature data). Passed
eventually to \code{\link[=are_equal]{are_equal()}}.}
}
\description{
Diff tool for the differences between two \code{soma_adat} objects.
When diffs of the table \emph{values} are interrogated, \strong{only}
the intersect of the column meta data or feature data is considered
}
\note{
Diffs of the column \emph{intersect} names are considered.
}
\examples{
diffAdats(example_data, example_data)

# remove random column
rm <- withr::with_seed(123, sample(1:ncol(example_data), 1))
diffAdats(example_data, example_data[, -rm])

# change Subarray randomly
diffAdats(example_data, dplyr::mutate(example_data, Subarray = sample(Subarray)))

# modify 2 RFUs randomly
new <- example_data
new[5, c(rm, rm + 1)] <- 999
diffAdats(example_data, new)
}
\seealso{
\code{\link[=are_equal]{are_equal()}}
}
\author{
Stu Field
}
