% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFeatureData.R
\name{getFeatureData}
\alias{getFeatureData}
\title{Get Feature Data (tibble)}
\usage{
getFeatureData(adat)
}
\arguments{
\item{adat}{An ADAT with intact attributes (i.e. has not been modified thus
stripping original attributes), typically created using \code{\link[=read_adat]{read_adat()}}.}
}
\value{
A \code{\link[=tibble]{tibble()}} with columns corresponding to the column meta
data entries in the ADAT. The tibble has a designated column "AptName"
corresponding to the features in the ADAT, which can be retrieved using
\code{\link[=getFeatures]{getFeatures()}}. This column can be use for indexing desired analytes.
}
\description{
Uses the Column Meta data (meta data that appears above the protein
measurements in the adat file) from the intact attributes of an ADAT and
compiles them into a \code{tibble} (\code{\link[=tibble]{tibble()}}) object for simple
manipulation and indexing. The Feature names of the ADAT become the
first column named \code{AptName}.
}
\examples{
# Attribute check
is.intact.attributes(sample.adat)   # must be TRUE

# Get feature table
tbl <- getFeatureData(sample.adat)
head(tbl)                    # First few rows of the data.frame
table(tbl$Dilution)          # Print number of features in each dilution.
choose5 <- sample(1:nrow(tbl), 5)  # Get 5 random rows (features)
tbl[ choose5, ]              # Print feature data for these 5
}
\seealso{
\code{\link[=getFeatures]{getFeatures()}}
}
\author{
Stu Field
}
