% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFeatures.R
\name{getFeatures}
\alias{getFeatures}
\title{Get Features}
\usage{
getFeatures(x, n = FALSE, rm.controls = FALSE)
}
\arguments{
\item{x}{Usually a \code{soma_adat} class object created using \code{\link[=read_adat]{read_adat()}}.}

\item{n}{Logical. Return an integer corresponding to the \emph{length}
of the returned string?}

\item{rm.controls}{Logical. Should all control and non-human analytes
(e.g. \code{HybControls}, \code{Non-Human}, \code{Non-Biotin}, \code{Spuriomer}) be removed
from the returned value?}
}
\value{
A character vector of ADAT feature ("analyte") names
or an integer number corresponding to the length of the
feature names (if \code{n = TRUE}).
}
\description{
Return the feature names (i.e. the column names for the feature data)
from a \code{soma_adat}, data frame, matrix, list, or character
vector. S3 methods exist for these classes.
}
\examples{
apts <- getFeatures(example_data)
head(apts)

getFeatures(example_data, TRUE)
getFeatures(example_data, n = TRUE)
bb <- getFeatures(names(example_data))
identical(apts, bb)

# Control Analytes
crtl <- setdiff(apts, getFeatures(example_data, rm.controls = TRUE))

getFeatureData(example_data) \%>\% filter(AptName \%in\% crtl)
}
\seealso{
\code{\link[=regex]{regex()}}, \code{\link[=grep]{grep()}}
}
\author{
Stu Field
}
