% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMeta.R
\name{getMeta}
\alias{getMeta}
\title{Get Meta Data Field Names}
\usage{
getMeta(x, n = FALSE)
}
\arguments{
\item{x}{Usually a \code{soma_adat} class object created using \code{\link[=read_adat]{read_adat()}}.}

\item{n}{Logical. Return an integer corresponding to the \emph{length}
of the returned string?}
}
\value{
A character vector of ADAT meta data names
or an integer number corresponding to the length of the
feature names (if \code{n = TRUE}).
}
\description{
Return a character vector of field names of the meta data
for an ADAT, or \code{soma.data} object.
}
\examples{
meta.vec <- getMeta(example_data)
head(meta.vec, 20)
getMeta(example_data, n = TRUE)

# test data.frame and character S3 methods
identical(getMeta(example_data), getMeta(names(example_data))) # TRUE
}
\author{
Stu Field
}
