% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MathGenerics.R
\name{MathGenerics}
\alias{MathGenerics}
\alias{Math.soma_adat}
\title{Mathematical Group Generics for ADAT Object}
\usage{
\method{Math}{soma_adat}(x, ...)
}
\arguments{
\item{x}{The \code{soma_adat} class object to perform the transformation.}

\item{...}{Additional arguments passed to the various group generics
as appropriate.}
}
\value{
A \code{soma_adat} object with the same dimensions of the input
object with the feature columns transformed by the specified generic.
}
\description{
This is the S3 group generic method to apply mathematical functions
to the RFU data of \code{soma_adat} objects.
The clinical meta data is \emph{not} transformed and remains in
the returned object. Typical generic functions include:
\itemize{
\item \code{log()}
\item \code{abs()}
\item \code{sign()}
\item \code{floor()}
\item \code{sqrt()}
\item \code{exp()}
\item See \code{\link[=groupGeneric]{groupGeneric()}} (\emph{Math}) for full listing
}
}
\examples{
example_data$seq.3343.1

# log-transformation
a <- log(example_data)
a$seq.3343.1
b <- log10(example_data)
b$seq.3343.1
isTRUE(all.equal(b, log(example_data, base = 10)))

# floor
c <- floor(example_data)
c$seq.3343.1

# square-root
d <- sqrt(example_data)
d$seq.3343.1

# rounding
e <- round(example_data)
e$seq.3343.1
}
\seealso{
\code{\link[=groupGeneric]{groupGeneric()}}
}
\author{
Stu Field
}
