% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivotExpressionSet.R
\name{pivotExpressionSet}
\alias{pivotExpressionSet}
\title{Convert to Long Format}
\usage{
pivotExpressionSet(eSet)
}
\arguments{
\item{eSet}{An \code{ExpressionSet} class object, created using \code{\link[=adat2eSet]{adat2eSet()}}.}
}
\value{
An \code{tibble} class consisting of the long format
conversion of an \code{ExpressionSet} object.
}
\description{
Utility function that converts an \code{"ExpressionSet"} object
from the "wide" data format to the "long" format \code{\link[=pivot_longer]{pivot_longer()}}.
}
\examples{
# subet into a reduced ADAT object
# 5 clinical variables and 3 features
# 10 samples
sub_adat <- sample.adat[1:10, c(1:5, 25:27)]
ex_set <- adat2eSet(sub_adat)

# convert to long format
adat_long <- pivotExpressionSet(ex_set)
}
\seealso{
\code{\link[=adat2eSet]{adat2eSet()}}, \pkg{Biobase}
}
\author{
Stu Field
}
