% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{package}
\name{SomaDataIO-deprecated}
\alias{SomaDataIO-deprecated}
\alias{meltExpressionSet}
\alias{getSomamers}
\alias{getSomamerData}
\title{Deprecated function(s) of the \code{SomaDataIO} package}
\arguments{
\item{...}{A simple pass-through to the replacement function.}
}
\description{
These functions are provided for compatibility with
older versions of the \code{SomaDataIO} package.
They may eventually be completely removed, so
please re-code your scripts accordingly as soon as possible.
}
\details{
\tabular{rl}{
\code{meltExpressionSet()} \tab now use \code{\link[=pivotExpressionSet]{pivotExpressionSet()}} \cr
\code{getSomamers()}       \tab now use \code{\link[=getFeatures]{getFeatures()}} \cr
\code{getSomamerData()}    \tab now use \code{\link[=getFeatureData]{getFeatureData()}}
}
}
\author{
Stu Field
}
